requirement.txt
1.django
2.sql